refactor: 코드 리뷰
- url mapping 전체 경로 명시
- 샘플 코드 deprecated 처리
- 테이블명 변경 - prefix 제거
- dto 내 request, response 분리
- 임베디드 레디스 버전 변경
- slf4j 의존성 충돌 문제로 한 버전 낮은것 사용
- 스프링 이벤트 상속 대신 애너테이션 사용
