--- ladder-api/src/main/kotlin/kr/mashup/ladder/room/dto/RoomDtos.kt
@@ -1,13 +1,0 @@
-package kr.mashup.ladder.room.dto
-
-data class RoomCreateRequest(
-    val description: String,
-)
-
-data class RoomChatRequest(
-    val chat: String,
-)
-
-data class RoomChatResponse(
-    val chat: String,
-)
