+++ ladder-api/src/main/kotlin/kr/mashup/ladder/room/dto/request/RoomChatRequest.kt
@@ -1,0 +1,5 @@
+package kr.mashup.ladder.room.dto.request
+
+data class RoomChatRequest(
+    val chat: String,
+)
