+++ ladder-api/src/main/kotlin/kr/mashup/ladder/room/dto/request/RoomCreateRequest.kt
@@ -1,0 +1,5 @@
+package kr.mashup.ladder.room.dto.request
+
+data class RoomCreateRequest(
+    val description: String,
+)
