+++ ladder-api/src/main/kotlin/kr/mashup/ladder/room/dto/response/RoomChatResponse.kt
@@ -1,0 +1,5 @@
+package kr.mashup.ladder.room.dto.response
+
+data class RoomChatResponse(
+    val chat: String,
+)
