+++ ladder-domain/src/main/kotlin/kr/mashup/ladder/domain/room/domain/RoomChatReceiveEvent.kt
@@ -1,9 +1,6 @@
 package kr.mashup.ladder.domain.room.domain
 
-import org.springframework.context.ApplicationEvent
-
 class RoomChatReceiveEvent(
-    source: Any,
     val roomId: Long,
-    val chat: String
-) : ApplicationEvent(source)
+    val chat: String,
+)
