+++ ladder-domain/src/main/kotlin/kr/mashup/ladder/domain/video/domain/VideoQueryRepository.kt
@@ -1,5 +1,6 @@
 package kr.mashup.ladder.domain.video.domain
 
+@Deprecated(message = "샘플 코드")
 interface VideoQueryRepository {
 
     fun findAll(): List<Video>
