+++ ladder-domain/src/main/kotlin/kr/mashup/ladder/domain/video/domain/VideoRepository.kt
@@ -1,5 +1,6 @@
 package kr.mashup.ladder.domain.video.domain
 
+@Deprecated(message = "샘플 코드")
 interface VideoRepository {
 
     fun save(video: Video): Video
